# src/main/java/com/aalonso/carregistry/dto/VehicleOrBrandIdRequest.java
package com.aalonso.carregistry.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class VehicleOrBrandIdRequest {
    private String id;
}