# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Vehicle Registry API
  description: Api to manage the vehicle registry
  contact:
    email: adanalonsomedin@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:8080/v1/car-registry-API
    description: Url to access the API
tags:
  - name: show_vehicles
    description: Check the vehicles of the database
  - name: add_vehicles
    description: Add new vehicles to the database
  - name: update_vehicles
    description: Update vehicles from the database
  - name: delete_vehicles
    description: Delete vehicles from the database
  - name: show_brands
    description: Check the vehicles of the database
  - name: add_brands
    description: Add new vehicles to the database
  - name: update_brands
    description: Update vehicles from the database
  - name: delete_brands
    description: Delete vehicles from the database
  - name: sign_up
    description: Add new users to the database
  - name: log_in
    description: Login to the application
paths:
  /car_registry/show_all_vehicles:
    get:
      tags:
        - show_vehicles
      summary: Get the list of vehicles
      description: Returns the list of all vehicles
      operationId: showAllVehicles
      security:
        - CarRegistry_auth: [ ]
      responses:
        '200':
          description: List found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
        '400':
          description: Bad request
        '404':
          description: Not vehicles found
        default:
          description: Unexpected error
        '500':
          description: Internal error
  /car_registry/show_vehicle_by_id:
    get:
      tags:
        - show_vehicles
      summary: Get vehicles
      description: Returns a vehicle by id
      operationId: showVehicleById
      security:
        - CarRegistry_auth: [ ]
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: The id of the vehicle to retrieve
      responses:
        '200':
          description: Vehicle found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          description: Bad request
        '404':
          description: Vehicle not found
        default:
          description: Unexpected error
        '500':
          description: Internal error
  /car_registry/show_all_brands:
    get:
      tags:
        - show_brands
      summary: Get the list of brands
      description: Returns the list of all brands
      operationId: showAllBrands
      security:
        - CarRegistry_auth: [ ]
      responses:
        '200':
          description: List found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Brand'
        '400':
          description: Bad request
        '404':
          description: Not brands found
        default:
          description: Unexpected error
        '500':
          description: Internal error
  /car_registry/show_brand_by_id:
    get:
      tags:
        - show_brands
      summary: Get brand
      description: Returns a brand by id
      operationId: showBrandById
      security:
        - CarRegistry_auth: [ ]
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: The id of the brand to retrieve
      responses:
        '200':
          description: Brand found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
        '400':
          description: Bad request
        '404':
          description: Brand not found
        default:
          description: Unexpected error
        '500':
          description: Internal error
  /car_registry/add_vehicles:
    post:
      tags:
        - add_vehicles
      summary: Add vehicles
      description: Add a vehicle list
      operationId: addVehicles
      security:
        - CarRegistry_auth: [ ]
      requestBody:
        description: List of vehicles to add
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Vehicle'
        required: true
      responses:
        '200':
          description: Vehicles added successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
        '400':
          description: Bad request
        default:
          description: Unexpected error
        '500':
          description: Internal error
  /car_registry/add_brands:
    post:
      tags:
        - add_brands
      summary: Add brands
      description: Add a brand list
      operationId: addBrands
      security:
        - CarRegistry_auth: [ ]
      requestBody:
        description: List of brands to add
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Brand'
        required: true
      responses:
        '200':
          description: Brands added successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Brand'
        '400':
          description: Bad request
        default:
          description: Unexpected error
        '500':
          description: Internal error
  /car_registry/update_vehicle:
    put:
      tags:
        - update_vehicles
      summary: Update a vehicle
      description: Updates a vehicle
      operationId: updateVehicle
      security:
        - CarRegistry_auth: [ ]
      requestBody:
        description: Vehicle data to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vehicle'
        required: true
      responses:
        '200':
          description: Vehicle updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          description: Bad request
        '404':
          description: Vehicle not found
        default:
          description: Unexpected error
        '500':
          description: Internal error
  /car_registry/update_brand:
    put:
      tags:
        - update_brands
      summary: Update a brand
      description: Updates a brand
      operationId: updateBrand
      security:
        - CarRegistry_auth: [ ]
      requestBody:
        description: Brand data to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Brand'
        required: true
      responses:
        '200':
          description: Brand updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
        '400':
          description: Bad request
        '404':
          description: Brand not found
        default:
          description: Unexpected error
        '500':
          description: Internal error
  /car_registry/delete_vehicle:
    delete:
      tags:
        - delete_vehicles
      summary: Delete a vehicle
      description: Deletes a vehicle by id
      operationId: deleteVehicle
      security:
        - CarRegistry_auth: [ ]
      requestBody:
        description: Vehicle data to delete
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleOrBrandIdRequest'
        required: true
      responses:
        '200':
          description: Vehicle deleted successfully
        '400':
          description: Bad request
        '404':
          description: Vehicle not found
        default:
          description: Unexpected error
        '500':
          description: Internal error
  /car_registry/delete_brand:
    delete:
      tags:
        - delete_brands
      summary: Delete a brand
      description: Delete a brand by id
      operationId: deleteBrand
      security:
        - CarRegistry_auth: [ ]
      requestBody:
        description: Brand data to delete
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleOrBrandIdRequest'
        required: true
      responses:
        '200':
          description: Brand deleted successfully
        '400':
          description: Bad request
        '404':
          description: Brand not found
        default:
          description: Unexpected error
        '500':
          description: Internal error
  /user/sign_up:
    post:
      tags:
        - sign_up
      summary: Add user
      description: Add a new user
      operationId: addNewUser
      requestBody:
        description: User to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: User added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        default:
          description: Unexpected error
        '500':
          description: Internal error
  /user/log_in:
    post:
      tags:
        - log_in
      summary: Log in
      description: Log in to the application
      operationId: logIn
      requestBody:
        description: User data to log in
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
              type: object
              properties:
                token:
                  type: string
        '400':
          description: Bad request
        default:
          description: Unexpected error
        '500':
          description: Internal error
components:
  schemas:
    Vehicle:
      type: object
      properties:
        id:
          type: string
        brand:
          $ref: '#/components/schemas/Brand'
        model:
          type: string
        mileage:
          type: integer
        price:
          type: number
          format: double
        year:
          type: integer
        colour:
          type: string
        fuelType:
          type: string
        numDoors:
          type: integer
        description:
          type: string
    Brand:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        warranty:
          type: integer
        country:
          type: string
    VehicleOrBrandIdRequest:
      type: object
      properties:
        id:
          type: string
    User:
      type: object
      properties:
        name:
          type: string
        password:
          type: string
        email:
          type: string
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
  securitySchemes:
    CarRegistry_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT